+++ packages/snap/src/handlers/cronjob.tsx
@@ -1,4 +1,5 @@
 import type { JsonRpcRequest } from '@metamask/snaps-sdk';
+import { type Types, utils } from 'tronweb';
 
 import type { PriceApiClient } from '../clients/price-api/PriceApiClient';
 import type { SnapClient } from '../clients/snap/SnapClient';
@@ -372,12 +373,16 @@ export class CronHandler {
     }
 
     // Skip if required fields are missing
-    if (!interfaceContext.fromAddress || !interfaceContext.scope) {
+    if (
+      !interfaceContext.fromAddress ||
+      !interfaceContext.scope ||
+      !interfaceContext.transactionRawData
+    ) {
       this.#logger.info('Context is missing required fields for scan refresh');
       return;
     }
 
-    const { preferences, scope, fromAddress, origin, scanParameters } =
+    const { preferences, scope, fromAddress, origin, transactionRawData } =
       interfaceContext;
 
     try {
@@ -413,12 +418,8 @@ export class CronHandler {
       try {
         scan = await this.#transactionScanService.scanTransaction({
           accountAddress: fromAddress,
-          parameters: {
-            from: scanParameters?.from ?? undefined,
-            to: scanParameters?.to ?? undefined,
-            data: scanParameters?.data ?? undefined,
-            value: scanParameters?.value ?? undefined,
-          },
+          transactionRawData:
+            transactionRawData as unknown as Types.Transaction['raw_data'],
           origin,
           scope,
           options,
@@ -552,7 +553,7 @@ export class CronHandler {
       return;
     }
 
-    const { preferences, scope, account, origin, scanParameters } =
+    const { preferences, scope, account, origin, transaction } =
       interfaceContext;
 
     // Determine what needs to be refreshed
@@ -588,15 +589,16 @@ export class CronHandler {
           options.push('validation');
         }
 
+        const transactionRawData: Types.Transaction['raw_data'] =
+          utils.deserializeTx.deserializeTransaction(
+            transaction.type,
+            transaction.rawDataHex,
+          );
+
         try {
           scan = await this.#transactionScanService.scanTransaction({
             accountAddress: account.address,
-            parameters: {
-              from: scanParameters?.from ?? undefined,
-              to: scanParameters?.to ?? undefined,
-              data: scanParameters?.data ?? undefined,
-              value: scanParameters?.value ?? undefined,
-            },
+            transactionRawData,
             origin,
             scope,
             options,