+++ packages/snap/CHANGELOG.md
@@ -36,6 +36,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Correct account discovery by using address-based activity checks instead of account IDs ([#206](https://github.com/MetaMask/snap-tron-wallet/pull/206))
 
+### Fixed
+
+- Improve simulation accuracy for contract transactions ([#186](https://github.com/MetaMask/snap-tron-wallet/pull/186))
+  - Supported transactions are those single-contract interaction transactions of the following types: `TransferContract`, `CreateSmartContract`, `TriggerSmartContract`.
+  - Unsupported transactions will show empty estimated changes and allow the user to proceed without blocking the confirmation.
+
 ## [1.22.0]
 
 ### Added